Vignette
일어났을 때 Vignette 초기화, 일부 버그 수정 및 리팩터링

+++ Unreal/Source/Unreal/Incubator.h
@@ -36,7 +36,7 @@ private:
 	EHabitat IncubatorHabitat;
 
 	UCommodity* GrowingCommodityRef;
-	float CommodityGrowthDuration;
+	int32 CommodityGrowthDuration;
 
 	FDateTime StartGrowingTime;
